==> 1_preparation_steps.ipynb <==
## ✫Livedoorテキストを使用するための準備（テキストの成形と各辞書の作成）

### 1. Livedoorテキストの読み込み

# # Livedoorニュースのファイルをダウンロード
# ! wget "https://www.rondhuit.com/download/ldcc-20140209.tar.gz"

# ファイルを解凍し、カテゴリー数と内容を確認
import tarfile
import os

# アーカイブの展開
tar = tarfile.open("ldcc-20140209.tar.gz", "r:gz")
tar.extractall("./data/livedoor/")
tar.close()

# フォルダのファイルとディレクトリを確認
files_folders = [name for name in os.listdir("./data/livedoor/text/")]
print(files_folders)

# カテゴリーのフォルダのみを抽出
categories = [name for name in os.listdir(
    "./data/livedoor/text/") if os.path.isdir("./data/livedoor/text/"+name)]

print("カテゴリー数:", len(categories))
print(categories)

# ['CHANGES.txt', 'dokujo-tsushin', 'it-life-hack', 'kaden-channel', 'livedoor-homme', 'movie-enter', 'peachy', 'README.txt', 'smax', 'sports-watch', 'topic-news']
# カテゴリー数: 9
# ['dokujo-tsushin', 'it-life-hack', 'kaden-channel', 'livedoor-homme', 'movie-enter', 'peachy', 'smax', 'sports-watch', 'topic-news']

# ファイルの中身を確認してみる
file_name = "./data/livedoor/text/movie-enter/movie-enter-6255260.txt"

with open(file_name) as text_file:
    text = text_file.readlines()
    print("0：", text[0])  # URL情報
    print("1：", text[1])  # タイムスタンプ
    print("2：", text[2])  # タイトル
    print("3：", text[3])  # 本文

# テキスト本文のみを取りだすための関数
def extract_main_txt(file_name):
    with open(file_name) as text_file:
        # 今回はタイトル行は外したいので、3要素目以降の本文のみ使用
        text = text_file.readlines()[3:]

        # 3要素目以降にも本文が入っている場合があるので、リストにして、後で結合させる
        text = [sentence.strip() for sentence in text]  # 空白文字(スペースやタブ、改行)の削除
        text = list(filter(lambda line: line != '', text))
        text = ''.join(text)
        text = text.translate(str.maketrans(
            {'\n': '', '\t': '', '\r': '', '\u3000': ''}))  # 改行やタブ、全角スペースを消す
        return text

# リストに前処理した本文と、カテゴリーのラベルを追加していく
import glob

list_text = []
list_label = []

for cat in categories:
    text_files = glob.glob(os.path.join("./data/livedoor/text", cat, "*.txt"))

    # 前処理extract_main_txtを実施して本文を取得
    body = [extract_main_txt(text_file) for text_file in text_files]

    label = [cat] * len(body)  # bodyの数文だけカテゴリー名のラベルのリストを作成

    list_text.extend(body)  # appendが要素を追加するのに対して、extendはリストごと追加する
    list_label.extend(label)

# 10番目の文章とラベルを確認
print(list_text[10])
print(list_label[10])

# pandasのDataFrameに変換する
import pandas as pd

df = pd.DataFrame({'text': list_text, 'label': list_label})

# 大きさを確認しておく（7,376文章が存在）
print(df.shape)
# (7376, 2)

df.head(3)

# カテゴリーの辞書を作成
dic_id2cat = dict(zip(list(range(len(categories))), categories))
dic_cat2id = dict(zip(categories, list(range(len(categories)))))

print(dic_id2cat)
print(dic_cat2id)

# DataFrameにカテゴリーindexの列を作成
df["label_index"] = df["label"].map(dic_cat2id)
df.head()

# label列を消去し、text, indexの順番にする
df = df.loc[:, ["text", "label_index"]]
df.head(3)

# 順番をシャッフルする
df = df.sample(frac=1, random_state=123).reset_index(drop=True)
df.head(3)

# # 件数が多いので2000件のサンプルを取得
df = df.sample(n=2000)
df.count()

### 2. Triplet Lossの処理をするための準備

# Series処理で読み込まれたラベルとDataframeのラベルを比較して同じもの（positive）を検索、サンプルとして１行だけ取り出す
def func(x):
    return df[['text']][df['label_index'] == x['label_index']].sample(1).iloc[0]

df = df.assign(positive = lambda x: x.apply(func, axis=1))

def func(x):
    return df[['label_index']][df['label_index'] == x['label_index']].sample(1).iloc[0]

df = df.assign(positive_index = lambda x: x.apply(func, axis=1))

# Series処理で読み込まれたラベルとDataframeのラベルを比較して異なるもの（negative）を検索、サンプルとして１行だけ取り出す
def func(x):
    return df[['text']][df['label_index'] != x['label_index']].sample(1).iloc[0]

df = df.assign(negative = lambda x: x.apply(func, axis=1))

def func(x):
    return df[['label_index']][df['label_index'] != x['label_index']].sample(1).iloc[0]

df = df.assign(negative_index = lambda x: x.apply(func, axis=1))

df.head(3)

# 不要な列の削除
df = df.drop(['label_index', 'positive_index', 'negative_index'], axis=1)

df.head(3)

# 列名の変更
df = df.rename(columns={'text': 'anchor'})

df.head(3)

# 件数の確認
df.shape
# (2000, 3)

from sklearn.model_selection import train_test_split
train, test = train_test_split(df, train_size=0.8, random_state=4)

print(len(train))
print(len(test))

# タブ区切りのテキストデータに書き出し
train.to_csv('./triplet_train.tsv', sep='\t', index = False)
test.to_csv('./triplet_test.tsv', sep='\t', index = False)
