==> 4_faiss_steps.ipynb <==
## ✫faissによる近似近傍探索処理

# !pip install faiss-gpu

# GPUによる処理を前提
import faiss

res = faiss.StandardGpuResources()
dim = 768
nlist = 256
M = 32
nbits = 8
metric = faiss.METRIC_L2
ivfpq_config = faiss.GpuIndexIVFPQConfig()
ivfpq_config.usePrecomputedTables = True
index = faiss.GpuIndexIVFPQ(res, dim, nlist, M, nbits, metric, ivfpq_config)

# 前のステップまでに作成した（文章ID, 文章ベクトル）を配列としたリストをロードする
import pickle

with open('livedoor_vec_list.pkl', 'rb') as f:
    livedoor_vec_list = pickle.load(f)

len(livedoor_vec_list)
# 7376

# faissによる学習データの準備
import numpy as np
import random
assert not index.is_trained
train_data = [v for _, v in livedoor_vec_list]
train_data = np.array(train_data, dtype=np.float32)

train_data.shape
# (7376, 768)

train_data = np.concatenate([train_data, train_data])

index.train(train_data)
assert index.is_trained

batch_size = 10000
for i in range(0, len(livedoor_vec_list), batch_size):
    input_vecs = []
    input_ids = []
    for item_id, vec in livedoor_vec_list[i:i+batch_size]:
        input_vecs.append(vec)
        input_ids.append(item_id)
    input_vecs = np.array(input_vecs, dtype=np.float32)
    input_ids = np.array(input_ids, dtype=np.int64)
    index.add_with_ids(input_vecs, input_ids)

# 学習済みインデックスの保存（GPUからCPUへ戻す）
converted_index_for_cpu = faiss.index_gpu_to_cpu(index)
faiss.write_index(converted_index_for_cpu, 'livedoor.index')

type(index)

### 本番利用時には保存されたindexからロードする処理から開始

import faiss
cpu_index = faiss.read_index('livedoor.index')
index = faiss.index_cpu_to_all_gpus(cpu_index)

# 文章ベクトル → 文章本文の辞書
import pickle

with open('livedoor_vec_to_text_dict.pkl', 'rb') as f:
    livedoor_vec_to_text_dict = pickle.load(f)

# 文章ベクトルリスト
import pickle
with open('livedoor_vec_list.pkl', 'rb') as f:
    livedoor_vec_list = pickle.load(f)

# 文章本文 → 文章ラベルの辞書
import pickle

with open('livedoor_text_to_label_dict.pkl', 'rb') as f:
    livedoor_text_to_label_dict = pickle.load(f)

### 検索用のベクトル変換

import pandas as pd
df = pd.read_csv('./triplet_test.tsv', sep='\t')
# df.head(3)
test_text = str(df['anchor'][6])

print(test_text)

import sentence_transformers
import numpy as np
from sentence_transformers import SentenceTransformer

sbert = SentenceTransformer('./sbert')
vector = sbert.encode(test_text)
# queries = np.array([vector], dtype=np.float32)
# queries.shape
# (1, 768)

import faiss
import numpy as np

vecs = np.array([vector], dtype=np.float32)
d, i = index.search(vecs, 10) # 近傍ベクトルを10個とってくる
d = d[0]
i = i[0]
for item_id, distance in zip(i, d):
    # 文章IDと距離を表示
    print('文章ID:', item_id, '距離:', distance)
    n = np.array(livedoor_vec_list[item_id][1], dtype=np.float32)
    n = tuple(n)
    
    # 文章IDから文章を参照
    print(livedoor_vec_to_text_dict[n])
    print(livedoor_text_to_label_dict[livedoor_vec_to_text_dict[n]])
    print('---')
